pyproject.toml
[project]
name = "sales-flow-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
